+++ clang-tools-extra/docs/ReleaseNotes.rst
@@ -203,6 +203,14 @@ New checks
   Finds virtual function overrides with different visibility than the function
   in the base class.
 
+- New :doc:`portability-avoid-platform-specific-fundamental-types
+  <clang-tidy/checks/portability/avoid-platform-specific-fundamental-types>` 
+  check.
+
+  Detects fundamental types (``int``, ``short``, ``long``, ``long long``,
+  ``char``, ``float``, etc) and warns against their use due to 
+  platform-dependent behavior.
+
 New check aliases
 ^^^^^^^^^^^^^^^^^